Rewrite queries in KdReportDao to use slick DSL. This avoids the postgreSQL problems with BOOLEAN columns since slick can handle the mapping itself for different DBs.
