+++ service_storagefacility/src/main/scala/no/uio/musit/microservice/storagefacility/dao/storage/OrganisationDao.scala
@@ -20,11 +20,10 @@
 package no.uio.musit.microservice.storagefacility.dao.storage
 
 import com.google.inject.{Inject, Singleton}
-import no.uio.musit.microservice.storagefacility.dao.SchemaName
 import no.uio.musit.microservice.storagefacility.domain.NodePath
 import no.uio.musit.microservice.storagefacility.domain.storage._
 import no.uio.musit.microservice.storagefacility.domain.storage.dto.{ExtendedStorageNode, OrganisationDto, StorageNodeDto}
-import no.uio.musit.service.MusitResults.{MusitDbError, MusitInternalError, MusitResult, MusitSuccess}
+import no.uio.musit.service.MusitResults.{MusitDbError, MusitResult, MusitSuccess}
 import play.api.Logger
 import play.api.db.slick.DatabaseConfigProvider
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
@@ -42,8 +41,6 @@ class OrganisationDao @Inject() (
   import driver.api._
 
   val logger = Logger(classOf[OrganisationDao])
-
-  private val organisationTable = TableQuery[OrganisationTable]
 
   private def updateAction(id: StorageNodeId, org: OrganisationDto): DBIO[Int] = {
     organisationTable.filter { ot =>
@@ -98,7 +95,8 @@ class OrganisationDao @Inject() (
 
   /**
    * Updates the path for the given StoragNodeId
-   * @param id the StorageNodeId to update
+   *
+   * @param id   the StorageNodeId to update
    * @param path the NodePath to set
    * @return MusitResult[Unit]
    */
@@ -127,25 +125,6 @@ class OrganisationDao @Inject() (
     }
 
     db.run(query.transactionally)
-  }
-
-  private class OrganisationTable(
-      val tag: Tag
-  ) extends Table[OrganisationDto](tag, SchemaName, "ORGANISATION") {
-
-    def * = (id, address) <> (create.tupled, destroy) // scalastyle:ignore
-
-    val id = column[Option[StorageNodeId]]("STORAGE_NODE_ID", O.PrimaryKey)
-    val address = column[Option[String]]("POSTAL_ADDRESS")
-
-    def create = (id: Option[StorageNodeId], address: Option[String]) =>
-      OrganisationDto(
-        id = id,
-        address = address
-      )
-
-    def destroy(organisation: OrganisationDto) =
-      Some((organisation.id, organisation.address))
   }
 
 }
