+++ service_storagefacility/src/main/scala/no/uio/musit/microservice/storagefacility/dao/storage/Shared.scala
@@ -20,7 +20,7 @@
 package no.uio.musit.microservice.storagefacility.dao.storage
 
 import no.uio.musit.microservice.storagefacility.dao._
-import no.uio.musit.microservice.storagefacility.domain.storage.dto.StorageUnitDto
+import no.uio.musit.microservice.storagefacility.domain.storage.dto.{BuildingDto, OrganisationDto, RoomDto, StorageUnitDto}
 import no.uio.musit.microservice.storagefacility.domain.storage.{StorageNodeId, StorageType}
 import no.uio.musit.microservice.storagefacility.domain.{NamedPathElement, NodePath}
 import play.api.Logger
@@ -43,6 +43,12 @@ private[dao] trait SharedStorageTables extends BaseStorageDao
   protected val rootNodeType: StorageType = StorageType.RootType
 
   protected val storageNodeTable = TableQuery[StorageNodeTable]
+
+  protected val organisationTable = TableQuery[OrganisationTable]
+
+  protected val buildingTable = TableQuery[BuildingTable]
+
+  protected val roomTable = TableQuery[RoomTable]
 
   /**
    * TODO: Document me!!!
@@ -228,6 +234,114 @@ private[dao] trait SharedStorageTables extends BaseStorageDao
         unit.isDeleted.getOrElse(false),
         unit.path
       ))
+  }
+
+  private[storage] class RoomTable(
+      val tag: Tag
+  ) extends Table[RoomDto](tag, SchemaName, "ROOM") {
+    // scalastyle:off method.name
+    def * = (
+      id,
+      perimeterSecurity,
+      theftProtection,
+      fireProtection,
+      waterDamage,
+      routinesAndContingency,
+      relativeHumidity,
+      temperatureAssessment,
+      lighting,
+      preventiveConservation
+    ) <> (create.tupled, destroy)
+
+    // scalastyle:on method.name
+
+    val id = column[Option[StorageNodeId]]("STORAGE_NODE_ID", O.PrimaryKey)
+    val perimeterSecurity = column[Option[Boolean]]("PERIMETER_SECURITY")
+    val theftProtection = column[Option[Boolean]]("THEFT_PROTECTION")
+    val fireProtection = column[Option[Boolean]]("FIRE_PROTECTION")
+    val waterDamage = column[Option[Boolean]]("WATER_DAMAGE_ASSESSMENT")
+    val routinesAndContingency = column[Option[Boolean]]("ROUTINES_AND_CONTINGENCY_PLAN")
+    val relativeHumidity = column[Option[Boolean]]("RELATIVE_HUMIDITY")
+    val temperatureAssessment = column[Option[Boolean]]("TEMPERATURE_ASSESSMENT")
+    val lighting = column[Option[Boolean]]("LIGHTING_CONDITION")
+    val preventiveConservation = column[Option[Boolean]]("PREVENTIVE_CONSERVATION")
+
+    def create = (
+      id: Option[StorageNodeId],
+      perimeterSecurity: Option[Boolean],
+      theftProtection: Option[Boolean],
+      fireProtection: Option[Boolean],
+      waterDamage: Option[Boolean],
+      routinesAndContingency: Option[Boolean],
+      relativeHumidity: Option[Boolean],
+      temperature: Option[Boolean],
+      lighting: Option[Boolean],
+      preventiveConservation: Option[Boolean]
+    ) =>
+      RoomDto(
+        id = id,
+        perimeterSecurity = perimeterSecurity,
+        theftProtection = theftProtection,
+        fireProtection = fireProtection,
+        waterDamageAssessment = waterDamage,
+        routinesAndContingencyPlan = routinesAndContingency,
+        relativeHumidity = relativeHumidity,
+        temperatureAssessment = temperature,
+        lightingCondition = lighting,
+        preventiveConservation = preventiveConservation
+      )
+
+    def destroy(room: RoomDto) =
+      Some((
+        room.id,
+        room.perimeterSecurity,
+        room.theftProtection,
+        room.fireProtection,
+        room.waterDamageAssessment,
+        room.routinesAndContingencyPlan,
+        room.relativeHumidity,
+        room.temperatureAssessment,
+        room.lightingCondition,
+        room.preventiveConservation
+      ))
+  }
+
+  private[storage] class BuildingTable(
+      val tag: Tag
+  ) extends Table[BuildingDto](tag, SchemaName, "BUILDING") {
+
+    def * = (id.?, address) <> (create.tupled, destroy) // scalastyle:ignore
+
+    val id = column[StorageNodeId]("STORAGE_NODE_ID", O.PrimaryKey)
+    val address = column[Option[String]]("POSTAL_ADDRESS")
+
+    def create = (id: Option[StorageNodeId], address: Option[String]) =>
+      BuildingDto(
+        id = id,
+        address = address
+      )
+
+    def destroy(building: BuildingDto) =
+      Some((building.id, building.address))
+  }
+
+  private[storage] class OrganisationTable(
+      val tag: Tag
+  ) extends Table[OrganisationDto](tag, SchemaName, "ORGANISATION") {
+
+    def * = (id, address) <> (create.tupled, destroy) // scalastyle:ignore
+
+    val id = column[Option[StorageNodeId]]("STORAGE_NODE_ID", O.PrimaryKey)
+    val address = column[Option[String]]("POSTAL_ADDRESS")
+
+    def create = (id: Option[StorageNodeId], address: Option[String]) =>
+      OrganisationDto(
+        id = id,
+        address = address
+      )
+
+    def destroy(organisation: OrganisationDto) =
+      Some((organisation.id, organisation.address))
   }
 
 }
